test: add scan business flow coverage
