+++ apps/server/src/app.test.ts
@@ -12,7 +12,7 @@ import { type Configuration, defaultConfiguration } from "@mdcz/shared/config";
 import { Website } from "@mdcz/shared/enums";
 import { afterEach, beforeEach, describe, expect, it } from "vitest";
 import { buildServer, type ServerApp } from "./app";
-import { closeTestServers, createTestServer, releaseTestServer } from "./app.testSupport";
+import { closeTestServers, createTestServer, releaseTestServer, syncMediaRootFromConfig } from "./app.testSupport";
 import type { RuntimeActionService } from "./services/runtimeActionService";
 import { formatSseEvent } from "./taskEvents";
 
@@ -29,31 +29,6 @@ const readStreamChunk = async (reader: ReadableStreamDefaultReader<Uint8Array>):
 };
 
 let standaloneServerApp: ServerApp | undefined;
-
-const syncMediaRootFromConfig = async (
-  fastify: ServerApp["fastify"],
-  token: string,
-  hostPath: string,
-): Promise<string> => {
-  await fastify.inject({
-    method: "POST",
-    url: "/trpc/config.update",
-    headers: { authorization: `Bearer ${token}` },
-    payload: { paths: { mediaPath: hostPath } },
-  });
-  const rootsResponse = await fastify.inject({
-    method: "GET",
-    url: "/trpc/mediaRoots.list",
-    headers: { authorization: `Bearer ${token}` },
-  });
-  const rootId = rootsResponse
-    .json()
-    .result.data.roots.find((rootDto: { hostPath: string }) => rootDto.hostPath === hostPath)?.id;
-  if (!rootId) {
-    throw new Error("Expected paths.mediaPath to create an enabled media root");
-  }
-  return rootId;
-};
 
 const startWebhookServer = async (): Promise<{
   close: () => Promise<void>;
@@ -782,222 +757,6 @@ describe("buildServer", () => {
     expect(rootResponse.json().result.data.entries.map((entry: { path: string }) => entry.path)).toContain(
       process.platform === "win32" ? root.replaceAll("\\", "/") : root,
     );
-  });
-
-  it("scans mounted media roots and serves persisted task details", async () => {
-    const root = await mkdtemp(join(tmpdir(), "mdcz-scan-root-"));
-    await mkdir(join(root, "nested"));
-    await writeFile(join(root, "nested", "movie.mp4"), "video");
-    await writeFile(join(root, "nested", "trailer.mp4"), "trailer");
-    await writeFile(join(root, "nested", "notes.txt"), "text");
-    const { fastify } = await createTestServer();
-    const loginResponse = await fastify.inject({
-      method: "POST",
-      url: "/trpc/auth.login",
-      payload: { password: "admin" },
-    });
-    const token = loginResponse.json().result.data.token;
-    await fastify.inject({
-      method: "POST",
-      url: "/trpc/config.update",
-      headers: { authorization: `Bearer ${token}` },
-      payload: { paths: { mediaPath: root } },
-    });
-    const rootsResponse = await fastify.inject({
-      method: "GET",
-      url: "/trpc/mediaRoots.list",
-      headers: { authorization: `Bearer ${token}` },
-    });
-    const rootId = rootsResponse
-      .json()
-      .result.data.roots.find((rootDto: { hostPath: string }) => rootDto.hostPath === root)?.id;
-    if (!rootId) {
-      throw new Error("Expected paths.mediaPath to create an enabled media root");
-    }
-
-    const startResponse = await fastify.inject({
-      method: "POST",
-      url: "/trpc/scans.start",
-      headers: { authorization: `Bearer ${token}` },
-      payload: { rootId },
-    });
-    const taskId = startResponse.json().result.data.id;
-
-    await expect
-      .poll(async () => {
-        const detailResponse = await fastify.inject({
-          method: "GET",
-          url: `/trpc/scans.detail?input=${encodeURIComponent(JSON.stringify({ taskId }))}`,
-          headers: { authorization: `Bearer ${token}` },
-        });
-        return detailResponse.json().result.data.task.status;
-      })
-      .toBe("completed");
-
-    const detailResponse = await fastify.inject({
-      method: "GET",
-      url: `/trpc/tasks.detail?input=${encodeURIComponent(JSON.stringify({ taskId }))}`,
-      headers: { authorization: `Bearer ${token}` },
-    });
-    const listResponse = await fastify.inject({
-      method: "GET",
-      url: "/trpc/tasks.list",
-      headers: { authorization: `Bearer ${token}` },
-    });
-    const libraryResponse = await fastify.inject({
-      method: "POST",
-      url: "/trpc/library.list",
-      headers: { authorization: `Bearer ${token}` },
-      payload: { query: "movie", limit: 20 },
-    });
-
-    const overviewResponse = await fastify.inject({
-      method: "GET",
-      url: "/trpc/overview.summary",
-      headers: { authorization: `Bearer ${token}` },
-    });
-    const logsResponse = await fastify.inject({
-      method: "GET",
-      url: "/trpc/logs.list",
-      headers: { authorization: `Bearer ${token}` },
-    });
-
-    expect(detailResponse.statusCode).toBe(200);
-    expect(listResponse.statusCode).toBe(200);
-    const rootDisplayName = root.split(/[\\/]+/u).at(-1);
-    expect(listResponse.json().result.data.tasks[0]).toMatchObject({
-      id: taskId,
-      kind: "scan",
-      rootDisplayName,
-    });
-    expect(detailResponse.json().result.data.task.videoCount).toBe(1);
-    expect(detailResponse.json().result.data.task.kind).toBe("scan");
-    expect(detailResponse.json().result.data.task.rootDisplayName).toBe(rootDisplayName);
-    expect(detailResponse.json().result.data.task.videos).toEqual(["nested/movie.mp4"]);
-    expect(libraryResponse.statusCode).toBe(200);
-    expect(libraryResponse.json().result.data).toEqual({ entries: [], total: 0 });
-    expect(overviewResponse.statusCode).toBe(200);
-    expect(overviewResponse.json().result.data.output).toMatchObject({ fileCount: 0, totalBytes: 0 });
-    expect(overviewResponse.json().result.data.recentAcquisitions).toEqual([]);
-    expect(detailResponse.json().result.data.events.map((event: { type: string }) => event.type)).toContain(
-      "completed",
-    );
-    expect(logsResponse.statusCode).toBe(200);
-    expect(logsResponse.json().result.data.logs).toContainEqual(
-      expect.objectContaining({ source: "task", type: "completed" }),
-    );
-    const clearLogsResponse = await fastify.inject({
-      method: "POST",
-      url: "/trpc/logs.clearRuntime",
-      headers: { authorization: `Bearer ${token}` },
-      payload: {},
-    });
-    const clearedLogsResponse = await fastify.inject({
-      method: "GET",
-      url: "/trpc/logs.list",
-      headers: { authorization: `Bearer ${token}` },
-    });
-    expect(clearLogsResponse.statusCode).toBe(200);
-    expect(clearedLogsResponse.json().result.data.logs).not.toContainEqual(
-      expect.objectContaining({ source: "task", type: "completed" }),
-    );
-    const retryResponse = await fastify.inject({
-      method: "POST",
-      url: "/trpc/tasks.retry",
-      headers: { authorization: `Bearer ${token}` },
-      payload: { taskId },
-    });
-    expect(retryResponse.statusCode).toBe(200);
-    expect(retryResponse.json().result.data.status).toBe("queued");
-
-    await expect
-      .poll(async () => {
-        const retriedDetailResponse = await fastify.inject({
-          method: "GET",
-          url: `/trpc/scans.detail?input=${encodeURIComponent(JSON.stringify({ taskId }))}`,
-          headers: { authorization: `Bearer ${token}` },
-        });
-        return retriedDetailResponse.json().result.data.task.status;
-      })
-      .toBe("completed");
-
-    const retriedLibraryResponse = await fastify.inject({
-      method: "POST",
-      url: "/trpc/library.list",
-      headers: { authorization: `Bearer ${token}` },
-      payload: { query: "movie", limit: 20 },
-    });
-    expect(retriedLibraryResponse.json().result.data.total).toBe(0);
-  });
-
-  it("lists ad-hoc scan candidates and filters non-media files", async () => {
-    const root = await mkdtemp(join(tmpdir(), "mdcz-candidates-root-"));
-    await mkdir(join(root, "nested"), { recursive: true });
-    await mkdir(join(root, "JAV_output"), { recursive: true });
-    await writeFile(join(root, "nested", "movie.mp4"), "video");
-    await writeFile(join(root, "nested", "trailer.mp4"), "trailer");
-    await writeFile(join(root, "nested", "notes.txt"), "text");
-    await writeFile(join(root, "JAV_output", "done.mp4"), "video");
-    const { fastify } = await createTestServer();
-    const loginResponse = await fastify.inject({
-      method: "POST",
-      url: "/trpc/auth.login",
-      payload: { password: "admin" },
-    });
-    const token = loginResponse.json().result.data.token;
-    const rootId = await syncMediaRootFromConfig(fastify, token, root);
-
-    const response = await fastify.inject({
-      method: "GET",
-      url: `/trpc/scans.candidates?input=${encodeURIComponent(
-        JSON.stringify({ scanDir: root, supportedExtensions: ["mp4"] }),
-      )}`,
-      headers: { authorization: `Bearer ${token}` },
-    });
-
-    expect(response.statusCode).toBe(200);
-    const candidates = response.json().result.data.candidates;
-    expect(candidates).toHaveLength(2);
-    expect(candidates).toEqual(
-      expect.arrayContaining([
-        expect.objectContaining({
-          name: "movie.mp4",
-          relativePath: "nested/movie.mp4",
-          rootId,
-          rootRelativePath: "nested/movie.mp4",
-          relativeDirectory: "nested",
-          extension: "mp4",
-        }),
-        expect.objectContaining({
-          name: "done.mp4",
-          relativePath: "JAV_output/done.mp4",
-          rootId,
-          rootRelativePath: "JAV_output/done.mp4",
-          relativeDirectory: "JAV_output",
-          extension: "mp4",
-        }),
-      ]),
-    );
-  });
-
-  it("returns no ad-hoc scan candidates for empty directories", async () => {
-    const root = await mkdtemp(join(tmpdir(), "mdcz-empty-candidates-root-"));
-    const { fastify } = await createTestServer();
-    const loginResponse = await fastify.inject({
-      method: "POST",
-      url: "/trpc/auth.login",
-      payload: { password: "admin" },
-    });
-    const token = loginResponse.json().result.data.token;
-
-    const response = await fastify.inject({
-      method: "GET",
-      url: `/trpc/scans.candidates?input=${encodeURIComponent(JSON.stringify({ scanDir: root }))}`,
-      headers: { authorization: `Bearer ${token}` },
-    });
-
-    expect(response.statusCode).toBe(200);
-    expect(response.json().result.data.candidates).toEqual([]);
   });
 
   it("protects automation REST endpoints and returns durable webhook payloads", async () => {
